Singleton-Decoder
Since the decoder holds the redemption-cache, responsible to be robust
against replay attacks, wrong and especially multiple-instantiation of
the decoder is an issue, e.g. if an application dev would create a new
decoder object for every token.

To reduce this latent security risk, the package creates an
attestationDecoder object, which can just be imported and hence is
guaranteed to only exist once per application.

For the unlikely case that caches really need to be cleared
programmatically, attestationDecoder.clearCaches() exists.